Cleaned up errors.
I cloned the repo to a new directory and rebuilt it.  There were some errors (due to strict mode I believe).  I cleaned those up.  Not sure why they didn't show in the other folder.  It was good to clone the repo and set it up again to make sure I can.