Added FileDesign that reads treatments from code.
I added FileDesign.js that will read the treatments from code.  I tried loading them from a CSV file like the .env file and that didn't work.  I will have to work on that.  Not sure if dotenv can be configured so that I can create a different dot file to commit to the repo to read treatments out of.  For now code will do.