Refactored code to introduce data classes.
Refactored code so that the questions and answer data is in it's own data classes.  The redux state contains a collection of QuestionAndAnswer classes with Question and Answer classes contained within it.
InitialState should be the same regardless of how the data is loaded, so I moved that back into questionSlice.js.
I renamed some of the global state